requirements.txt
elasticsearch==5.4.0
elasticsearch-dsl==5.3.0
influxdb==4.1.1
kafka-python==1.3.3
kazoo==2.4.0
lxml==3.8.0
MarkupSafe==1.0
matplotlib==2.0.2
numpy==1.13.1
pandas==0.20.3
pymongo==3.5.1
pytz==2017.2
pyzmq==16.0.2
redis==2.10.5
requests==2.17.3
